Final1

+++ src/Bounds.java
@@ -16,12 +16,12 @@ public class Bounds extends GameObject{
     public void render(Graphics g) {
         Graphics2D g2d = (Graphics2D) g;
         g.setColor(Color.blue);
-        g.fillRect(x,y,20,20);
+        g.fillRect((int)x,(int)y,20,20);
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 20, 20);
+        return new Rectangle((int)x, (int)y, 20, 20);
     }
 
 }
