+++ src/GameObject.java
@@ -2,11 +2,11 @@ import java.awt.*;
 
 public abstract class  GameObject {
 
-    protected int x, y;
+    protected float x, y;
     protected ID id;
     protected float velX = 0, velY = 0;
 
-    public GameObject(int x, int y, ID id) {
+    public GameObject(float x, float y, ID id) {
         this.x = x;
         this.y = y;
         this.id=id;
@@ -26,19 +26,19 @@ public abstract class  GameObject {
         this.id = id;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
@@ -46,7 +46,7 @@ public abstract class  GameObject {
         return velX;
     }
 
-    public void setVelX(int velX) {
+    public void setVelX(float velX) {
         this.velX = velX;
     }
 
@@ -54,7 +54,7 @@ public abstract class  GameObject {
         return velY;
     }
 
-    public void setVelY(int velY) {
+    public void setVelY(float velY) {
         this.velY = velY;
     }
