+++ src/Leaf.java
@@ -4,11 +4,13 @@ public class Leaf extends GameObject{
 
     private Handler handler;
     private PlayerInventory playerInventory;
+
     public Leaf(int x, int y, ID id, Handler handler, PlayerInventory playerInventory) {
         super(x, y, id);
 
         this.handler=handler;
         this.playerInventory=playerInventory;
+
     }
 
     @Override
@@ -20,12 +22,12 @@ public class Leaf extends GameObject{
     public void render(Graphics g) {
         Graphics2D g2d = (Graphics2D) g;
         g.setColor(Color.green);
-        g.fillRect(x,y,20,20);
+        g.fillRect((int)x,(int)y,20,20);
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x,y,30,30);
+        return new Rectangle((int)x,(int)y,30,30);
     }
 
     public void collision() {
@@ -37,6 +39,7 @@ public class Leaf extends GameObject{
                     this.playerInventory.increaseLeafCount();
                     handler.removeObject(this);
                     System.out.println("Leaf: "+this.playerInventory.getLeafCount());
+
 
                 }
             }
